requirements.txt
annotated-types==0.6.0
anyio==4.2.0
autoflake==2.2.1
bcrypt==4.1.2
black==24.1.1
boto3==1.34.39
botocore==1.34.39
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.1
cryptography==42.0.2
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.2
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jose==1.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
MouseInfo==0.1.3
mypy==1.8.0
mypy-extensions==1.0.0
orjson==3.9.13
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
psycopg2==2.9.9
pyasn1==0.5.1
PyAutoGUI==0.9.54
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.2
pyflakes==3.2.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytest==8.0.0
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.7
pytweening==1.0.7
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.0.7
uvicorn==0.27.0.post1
watchfiles==0.21.0
websockets==12.0
